+++ src/main.ts
@@ -1,3 +1,4 @@
+import mapboxgl from 'mapbox-gl';
 import { Plugin } from 'obsidian';
 import { changeApiTokenCommand } from './commands/api-token.command';
 import { changeDefaultMarkerUrlCommand } from './commands/custom-marker.command';
@@ -9,6 +10,7 @@ import { addNewLocationFromClipboard } from './commands/new-from-clipboard.comma
 import { toggleReverseCoordinates } from './commands/toggle-reverse.command';
 import { checkVersionUpdate } from './functions/version-hint.func';
 import { processLocationCodeBlock } from './processors/process-code.func';
+import { processInteractiveLocationCodeBlock } from './processors/process-interactive-code.func';
 import { processMultiLocationCodeBlock } from './processors/process-multi-locations.func';
 import { LocationSettingTab } from './settings/plugin-settings.tab';
 import { DEFAULT_SETTINGS, LocationPluginSettings } from './settings/plugin-settings.types';
@@ -31,6 +33,8 @@ export default class MapboxPlugin extends Plugin {
 		changeDefaultMarkerUrlCommand(this);
 		toggleReverseCoordinates(this);
 
+		mapboxgl.accessToken = this.settings.mapboxToken;
+
 		// Register the processors for the given code blocks.
 		// Code blocks are used in this plugin to render the maps.
 		this.registerMarkdownCodeBlockProcessor('location', (source: string, el: HTMLElement) =>
@@ -38,6 +42,11 @@ export default class MapboxPlugin extends Plugin {
 		);
 		this.registerMarkdownCodeBlockProcessor('multi-location', (source: string, el: HTMLElement) =>
 			processMultiLocationCodeBlock(source, el, this.settings),
+		);
+		this.registerMarkdownCodeBlockProcessor(
+			'interactive-location',
+			(source: string, el: HTMLElement) =>
+				processInteractiveLocationCodeBlock(source, el, this.settings),
 		);
 
 		// Adding the UI settings tab to the Obsidian preferences.