+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Next] - 2025-01-24
+
+### Changed
+
+- [⚙️] Removed some unused parameters
+
 ## [Next] - 2024-11-21
 
 ### Added
@@ -16,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [🚀] Added command support for default map style, marker color, custom marker URL and API token update
+- [🚀] Added first PoC support for interactive maps (and single marker)
 
 ## [Next] - 2024-10-23